Tira ranking fragment do bottomNavigation e coloca CourseUnit Fragment como principal
