+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -1,12 +1,12 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
 
-    <item
+    <!-- <item
         android:id="@+id/navigation_ranking"
         android:icon="@drawable/ranking_icon"
-        android:title="Ranking" />
+        android:title="Ranking" /> -->
     <item
-        android:id="@+id/navigation_questionarios"
+        android:id="@+id/navigation_course_unit"
         android:icon="@drawable/diciplinas_icon"
         android:title="Diciplinas" />
     <item
